refactor: props를 묶어서 정리
