+++ frontend/_apps/schedule/components/ScheduleEdit/ScheduleEditData.ts
@@ -1,68 +1,12 @@
 import CREATE from '@SCH/constants/CREATE';
+import { useGetCategory, usePostCategory } from '@SCH/hooks/useCategory';
 import { stringToNumber } from '@SCH/utils/formatDate';
-import { CategoryReadResponse, FriendReadResponse, UserReadResponse } from 'amadda-global-types';
+import { FriendReadResponse, UserReadResponse, ScheduleCreateRequest } from 'amadda-global-types';
 import { useRouter } from 'next/router';
-
 import type { Dispatch, ReactNode, SetStateAction } from 'react';
 import { useEffect, useState } from 'react';
+import { ScheduleEditDataProps } from './ScheduleEditDataProps';
 
-export interface ScheduleEditDataProps {
-  data: {
-    startYear: number | undefined;
-    setStartYear: Dispatch<SetStateAction<number | undefined>>;
-    startMonth: number | undefined;
-    setStartMonth: Dispatch<SetStateAction<number | undefined>>;
-    startDate: number | undefined;
-    setStartDate: Dispatch<SetStateAction<number | undefined>>;
-    startTime: number | undefined;
-    setStartTime: Dispatch<SetStateAction<number | undefined>>;
-    startMinute: number | undefined;
-    setStartMinute: Dispatch<SetStateAction<number | undefined>>;
-    endYear: number | undefined;
-    setEndYear: Dispatch<SetStateAction<number | undefined>>;
-    endMonth: number | undefined;
-    setEndMonth: Dispatch<SetStateAction<number | undefined>>;
-    endDate: number | undefined;
-    setEndDate: Dispatch<SetStateAction<number | undefined>>;
-    endTime: number | undefined;
-    setEndTime: Dispatch<SetStateAction<number | undefined>>;
-    endMinute: number | undefined;
-    setEndMinute: Dispatch<SetStateAction<number | undefined>>;
-    category: Array<CategoryReadResponse>;
-    scheduleName: string;
-    setScheduleName: Dispatch<SetStateAction<string>>;
-    participants: Array<UserReadResponse>;
-    setParticipants: Dispatch<SetStateAction<Array<UserReadResponse>>>;
-    scheduleStartAt: string;
-    setScheduleStartAt: Dispatch<SetStateAction<string>>;
-    scheduleEndAt: string;
-    setScheduleEndAt: Dispatch<SetStateAction<string>>;
-    isDateSelected: boolean;
-    setIsDateSelected: Dispatch<SetStateAction<boolean>>;
-    isTimeSelected: boolean;
-    setIsTimeSelected: Dispatch<SetStateAction<boolean>>;
-    isAllday: boolean;
-    setIsAllday: Dispatch<SetStateAction<boolean>>;
-    isAuthorizedAll: boolean;
-    setIsAuthorizedAll: Dispatch<SetStateAction<boolean>>;
-    scheduleContent: string;
-    setScheduleContent: Dispatch<SetStateAction<string>>;
-    alarmTime: keyof typeof CREATE.ALARMS | string;
-    setAlarmTime: Dispatch<SetStateAction<keyof typeof CREATE.ALARMS | string>>;
-    categorySeq?: number;
-    setCategorySeq: Dispatch<SetStateAction<number>>;
-    scheduleMemo: string;
-    setScheduleMemo: Dispatch<SetStateAction<string>>;
-    partyAutoComplete: (key: string) => void;
-    partySearchInput: string;
-    partySearchResult: FriendReadResponse;
-    categoryInput: string;
-    setCategoryInput: Dispatch<SetStateAction<string>>;
-    addCategory: (key: string) => void;
-    onDelete: (userSeq: number) => void;
-    submit: () => void;
-  };
-}
 export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps) => ReactNode }) {
   const router = useRouter();
 
@@ -95,22 +39,23 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
   const [categorySeq, setCategorySeq] = useState<number | undefined>(undefined);
   const [scheduleMemo, setScheduleMemo] = useState<string>('');
 
-  const [category, setCategory] = useState<Array<CategoryReadResponse>>([]);
-
   const [categoryInput, setCategoryInput] = useState<string>('');
 
-  const getCategory = () =>
-    fetch(`${process.env.NEXT_PUBLIC_SCHEDULE}/api/schedule/user/category`, {
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-      .then(res => res.json())
-      .then(json => setCategory(json.data));
-
-  useEffect(() => {
-    getCategory();
-  }, [category]);
+  const scheduleInputData = {
+    scheduleName: '',
+    participants: [],
+    scheduleStartAt: '',
+    scheduleEndAt: '',
+    isDateSelected: false,
+    isTimeSelected: false,
+    isAllday: false,
+    isAuthorizedAll: false,
+    scheduleContent: '',
+    alarmTime: 'NONE',
+    scheduleMemo: '',
+  };
+  const [writtenSchedule, setWrittenSchedule] = useState<ScheduleCreateRequest>(scheduleInputData);
+  const { category, categoryIsLoading } = useGetCategory();
 
   if (router.basePath.includes('edit')) {
     const { scheduleSeq } = router.query;
@@ -152,6 +97,7 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
   }
 
   const partyAutoComplete = async (key: string) => {
+    //TODO: SWR Migration
     setPartySearchInput(key);
     return fetch(`${process.env.NEXT_PUBLIC_USER}/api/friend?searchKey=${key}`, {
       headers: {
@@ -162,18 +108,7 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
       .then(json => setPartySearchResult(json.data));
   };
 
-  const addCategory = async (key: string) => {
-    fetch(`${process.env.NEXT_PUBLIC_USER}/api/category`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-      .then(res => res.json())
-      .then(json => setPartySearchResult(json.data))
-      .then(data => getCategory())
-      .catch(err => err);
-  };
+  const postCategory = (categoryName: string) => usePostCategory(categoryName);
 
   const onDelete = (userSeq: number) => {
     const result = participants.filter(p => p.userSeq !== userSeq);
@@ -181,26 +116,12 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
   };
 
   const submit = async () => {
-    let data;
     if (router.basePath.includes('edit')) {
       const { scheduleSeq } = router.query;
-      data = {
-        scheduleContent,
-        isDateSelected,
-        isTimeSelected,
-        isAllday,
-        isAuthorizedAll,
-        scheduleStartAt,
-        scheduleEndAt,
-        participants,
-        alarmTime,
-        scheduleName,
-        scheduleMemo,
-        categorySeq,
-      };
+
       return fetch(`${process.env.NEXT_PUBLIC_SCHEDULE}/api/schedule/${scheduleSeq}`, {
         method: 'PUT',
-        body: data,
+        body: JSON.stringify(writtenSchedule),
         headers: {
           'Content-Type': 'application/json',
         },
@@ -209,22 +130,9 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
         .then(json => router.push(`${process.env.NEXT_PUBLIC_SHELL}/schedule/${json.data}`))
         .catch(err => router.push(`${process.env.NEXT_PUBLIC_SHELL}/schedule`));
     }
-    data = {
-      scheduleContent,
-      isDateSelected,
-      isTimeSelected,
-      isAllday,
-      isAuthorizedAll,
-      scheduleStartAt,
-      scheduleEndAt,
-      participants,
-      alarmTime,
-      scheduleName,
-      scheduleMemo,
-    };
     return fetch(`${process.env.NEXT_PUBLIC_SCHEDULE}/api/schedule`, {
       method: 'POST',
-      body: data,
+      body: JSON.stringify(writtenSchedule),
       headers: {
         'Content-Type': 'application/json',
       },
@@ -234,12 +142,35 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
       .catch(err => router.push(`${process.env.NEXT_PUBLIC_SHELL}/schedule`));
   };
 
-  const data = {
+  const states = {
     startTime,
     startMonth,
     startDate,
     startMinute,
     startYear,
+    endYear,
+    endMinute,
+    endMonth,
+    endTime,
+    endDate,
+    category,
+    scheduleName,
+    participants,
+    scheduleStartAt,
+    scheduleEndAt,
+    isDateSelected,
+    isTimeSelected,
+    isAllday,
+    isAuthorizedAll,
+    scheduleContent,
+    alarmTime,
+    categorySeq,
+    scheduleMemo,
+    partySearchInput,
+    partySearchResult,
+    categoryInput,
+  };
+  const setStates = {
     setStartDate,
     setStartMinute,
     setStartMonth,
@@ -250,46 +181,29 @@ export function ScheduleEditData(props: { children: (args: ScheduleEditDataProps
     setEndMinute,
     setEndMonth,
     setEndYear,
-    endYear,
-    endMinute,
-    endMonth,
-    endTime,
-    endDate,
-    category,
-    scheduleName,
+
     setScheduleName,
-    participants,
     setParticipants,
-    scheduleStartAt,
     setScheduleStartAt,
-    scheduleEndAt,
     setScheduleEndAt,
-    isDateSelected,
     setIsDateSelected,
-    isTimeSelected,
     setIsTimeSelected,
-    isAllday,
     setIsAllday,
-    isAuthorizedAll,
     setIsAuthorizedAll,
-    scheduleContent,
     setScheduleContent,
-    alarmTime,
     setAlarmTime,
-    categorySeq,
     setCategorySeq,
-    scheduleMemo,
     setScheduleMemo,
-    partyAutoComplete,
-    partySearchInput,
-    partySearchResult,
-    categoryInput,
+
     setCategoryInput,
-    addCategory,
+  };
+  const fn = {
+    partyAutoComplete,
+    postCategory,
     onDelete,
     submit,
   };
   if (!props.children) return null;
 
-  return props.children({ data });
+  return props.children({ states, setStates, fn });
 }
